basic reworking of model to match formatting
